pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qinter"
dynamic = ["version"]
description = "Human-readable Python error explanations with beautiful CLI"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["debugging", "errors", "exceptions", "cli", "developer-tools"]
requires-python = ">=3.8"
dependencies = [
    "rich>=13.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/qinter"
Repository = "https://github.com/yourusername/qinter"
Documentation = "https://qinter.readthedocs.io"
"Bug Tracker" = "https://github.com/yourusername/qinter/issues"

[project.scripts]
qinter = "qinter.cli.main:main"

[tool.setuptools.dynamic]
version = {attr = "qinter.__version__.__version__"}

[tool.setuptools.packages.find]
include = ["qinter*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=qinter --cov-report=html --cov-report=term-missing"